Revert changes (#3481)

+++ packages/lexical/src/LexicalGC.ts
@@ -40,31 +40,37 @@ export function $garbageCollectDetachedDecorators(
 
 function $garbageCollectDetachedDeepChildNodes(
   node: ElementNode,
+  parentKey: NodeKey,
   prevNodeMap: NodeMap,
   nodeMap: NodeMap,
   dirtyNodes: Map<NodeKey, IntentionallyMarkedAsDirtyElement>,
 ): void {
-  let child = node.getFirstChild();
+  const children = node.__children;
+  const childrenLength = children.length;
 
-  while (child !== null) {
-    const childKey = child.__key;
-    const nextSibling = child.getNextSibling();
-    if ($isElementNode(child)) {
-      $garbageCollectDetachedDeepChildNodes(
-        child,
-        prevNodeMap,
-        nodeMap,
-        dirtyNodes,
-      );
-    }
+  for (let i = 0; i < childrenLength; i++) {
+    const childKey = children[i];
+    const child = nodeMap.get(childKey);
 
-    // If we have created a node and it was dereferenced, then also
-    // remove it from out dirty nodes Set.
-    if (!prevNodeMap.has(childKey)) {
-      dirtyNodes.delete(childKey);
+    if (child !== undefined && child.__parent === parentKey) {
+      if ($isElementNode(child)) {
+        $garbageCollectDetachedDeepChildNodes(
+          child,
+          childKey,
+          prevNodeMap,
+          nodeMap,
+          dirtyNodes,
+        );
+      }
+
+      // If we have created a node and it was dereferenced, then also
+      // remove it from out dirty nodes Set.
+      if (!prevNodeMap.has(childKey)) {
+        dirtyNodes.delete(childKey);
+      }
+
+      nodeMap.delete(childKey);
     }
-    nodeMap.delete(childKey);
-    child = nextSibling;
   }
 }
 
@@ -98,6 +104,7 @@ export function $garbageCollectDetachedNodes(
         if ($isElementNode(node)) {
           $garbageCollectDetachedDeepChildNodes(
             node,
+            nodeKey,
             prevNodeMap,
             nodeMap,
             dirtyElements,
